+++ admin/main_admin_check.php
@@ -6,15 +6,73 @@
  */
 
 
-require_once('wp-load.php' );
-require_once( 'wp-admin/includes/admin.php');
+// require_once('wp-load.php' );
+// require_once( 'wp-admin/includes/admin.php');
 
-$title=$_POST['title'];
-$excerpt=$_POST['excerpt'];
-$content=$_POST['content'];
-$category=$_POST['category'];
+// $title=$_POST['title'];
+// $excerpt=$_POST['excerpt'];
+// $content=$_POST['content'];
+// $category=$_POST['category'];
 
-var_dump($title);
+// var_dump($title);
+
+
+require_once( get_theme_file_path('processing.php') );
+
+$user_data=[
+    'mail' => $_POST['mail'],
+    'status' => $_POST['status'],
+    'pay_choice' => $_POST['pay_choice'],
+    'date' => date("Y-m-d H:i:s"),
+];
+json_encode($user_data);
+// обьект с данными хранится в tokens, а после регистрации и проверки токена сразу же добавляются поля из обьекта
+if ($result = sendRegistrationLink($email, $user_data)) {
+    echo $result;
+} else {
+    echo 'Failed to send the registration link.';
+}
+
+function sendRegistrationLink($email, $user_data) {
+    require_once( get_theme_file_path('send_mail.php') );
+    $registrationPage = 'https://nezhno.space/registration';
+    $token = generateUniqueToken();
+    storeToken( $email, $token, $user_data );
+
+    $registrationLink = $registrationPage . '?token=' . urlencode($token);
+    $subject = 'Registration Link';
+    $message = 'Please click on the following link to register: ' . $registrationLink;
+    $result = SendMail($email, $subject, $message,$subject);
+    if ($result) {
+        return $result; // Email sent successfully
+    } else {
+        return false; // Failed to send email
+    }
+}
+
+function generateUniqueToken($length = 10) {
+    $characters = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789';
+    $token = '';
+    $charactersLength = strlen($characters);
+    for ($i = 0; $i < $length; $i++) {
+        $randomIndex = mt_rand(0, $charactersLength - 1);
+        $token .= $characters[$randomIndex];
+    }
+    return $token;
+}
+
+function storeToken($email, $token, $user_data) {
+    require_once(get_theme_file_path('processing.php'));
+    $db = new SafeMySQL();
+    $query = "INSERT INTO tokens (mail, token, info) VALUES ('$email', '$token', '$user_data')";
+    if ($db->query($query)) {
+        return TRUE;
+    } else {
+        return FALSE;
+    }
+}
+
+
